+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## HEAD (2023-02-??)
+## 1.0.1 (2023-03-16)
 
 Fixed name of the attribute for the "nothing selected" option: `nonekey` has
 been renamed to `none_key` and `nonelabel` has been renamed to `none_label`.
@@ -9,6 +9,8 @@ Added the class `MenuItem` and `Panel` and the `App` attributes `menus` and
 `panel` for the menus/panels that appear on various LivingApps pages.
 
 Added attribute `App.datasource`.
+
+Fixed name of "geo distance" method (`Globals.dist()`).
 
 
 ## 1.0.0 (2023-02-02)